requirements.txt
Django==1.11
django-taggit==0.22.0
djangorestframework==3.6.2
pytz==2017.2
